Specify the string type for variables
Previously string was the default type, but this is no longer the
case:
https://www.terraform.io/docs/configuration/variables.html#type-constraints

I ran into a case today where this mattered.  I used
set([var.string_value]) in a for_each, and Terraform rejected the
syntax since it couldn't verify that the output would be a set of
strings.  Adding the string type declaration to the variable fixed the
issue.